Added devcontainer, WIP readme and conftest.py moved
