+++ .gitignore
@@ -6,10 +6,8 @@ __pycache__/
 *.py[cod]
 *$py.class
 
-# Dev Container
-**/.devcontainer
-
 # Test data tar gz
+**/cumulus-geoproc-test-data/
 *.tar.gz
 
 # C extensions
