--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "cumulus-geoproc-test-data"]
-	path = cumulus-geoproc-test-data
-	url = https://github.com/USACE/cumulus-geoproc-test-data
